requirements.txt
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.5
pandas==1.1.5
pdfminer.six==20200517
pdfplumber==0.5.27
Pillow==8.1.2
pycryptodome==3.10.1
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
six==1.15.0
sortedcontainers==2.3.0
Wand==0.6.6
Werkzeug==1.0.1
